+++ HISTORY.rst
@@ -25,6 +25,7 @@ latest (unreleased)
 
 * ALCYN-2385: set zetes user when confirmation confirmed by portal user
 * Disable copying esb_exported by default
+* ALCYN-2366: fix delivered qty assigned to consignment location
 * ALCYN-2364:Fix to add delivery fees on the SO before creating the invoice
 * ALCYN-2397: Display on customer statement on report, no supplier